split app components
